chore: rename project to @imgix/vue
