+++ .all-contributorsrc
@@ -18,7 +18,7 @@
     }
   ],
   "contributorsPerLine": 7,
-  "projectName": "vue-imgix",
+  "projectName": "vue",
   "projectOwner": "imgix",
   "repoType": "github",
   "repoHost": "https://github.com",
