+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -5,7 +5,7 @@ about: Suggest an idea for this project
 
 **Before you submit:**
 
-- [ ] Please search through the existing issues (both open AND closed) to see if your feature has already been discussed. Github issue search can be used for this: https://github.com/imgix/vue-imgix/issues?utf8=%E2%9C%93&q=is%3Aissue
+- [ ] Please search through the existing issues (both open AND closed) to see if your feature has already been discussed. Github issue search can be used for this: https://github.com/imgix/vue/issues?utf8=%E2%9C%93&q=is%3Aissue
 - [ ] Please take a moment to find out whether your idea fits with the scope and aims of the project
 
 **Is your feature request related to a problem? Please describe.**
