+++ .github/ISSUE_TEMPLATE/question.md
@@ -5,7 +5,7 @@ about: Ask a question about the project
 
 **Before you submit:**
 
-- [ ] Please search through the existing issues (both open AND closed) to see if your question has already been discussed. Github issue search can be used for this: https://github.com/imgix/vue-imgix/issues?utf8=%E2%9C%93&q=is%3Aissue
+- [ ] Please search through the existing issues (both open AND closed) to see if your question has already been discussed. Github issue search can be used for this: https://github.com/imgix/vue/issues?utf8=%E2%9C%93&q=is%3Aissue
 
 **Question**
 A clear and concise description of your question
