+++ public/index.html
@@ -5,12 +5,12 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width,initial-scale=1.0" />
     <link rel="icon" href="<%= BASE_URL %>favicon.ico" />
-    <title>vue-imgix Development Environment</title>
+    <title>@imgix/vue Development Environment</title>
   </head>
   <body>
     <noscript>
       <strong
-        >We're sorry but vue-imgix doesn't work properly without JavaScript
+        >We're sorry but @imgix/vue doesn't work properly without JavaScript
         enabled. Please enable it to continue.</strong
       >
     </noscript>
