+++ src/components/advanced/buildUrl.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script>
-import { buildUrl } from '@/plugins/vue-imgix';
+import { buildUrl } from '@/plugins/imgix-vue';
 
 // NB: Make sure initVueImgix has been called before this code runs
 export default {
