+++ src/plugins/imgix-vue/types.ts
