+++ tests/unit/initialization.spec.ts
@@ -4,15 +4,15 @@ describe('Initialization', () => {
   beforeEach(() => {
     jest.resetModules();
   });
-  it('allows the developer to initialise the vue-imgix library', () => {
+  it('allows the developer to initialise the @imgix/vue library', () => {
     const Vue = require('vue');
-    const VueImgix = require('@/plugins/vue-imgix');
+    const VueImgix = require('@/plugins/imgix-vue');
     Vue.use(VueImgix, {
       domain: 'test-domain.imgix.net',
     });
   });
-  it('throws an error if a method is used before vue-imgix was initialised', () => {
-    const { buildUrlObject } = require('@/plugins/vue-imgix');
+  it('throws an error if a method is used before @imgix/vue was initialised', () => {
+    const { buildUrlObject } = require('@/plugins/imgix-vue');
     expect(() => {
       buildUrlObject('/test-image.jpg');
     }).toThrow(/Vue\.use/);
@@ -20,13 +20,13 @@ describe('Initialization', () => {
 
   it(`doesn't include ixlib in generated urls when includeLibraryParam is set to false`, () => {
     const Vue = require('vue');
-    const VueImgix = require('@/plugins/vue-imgix');
+    const VueImgix = require('@/plugins/imgix-vue');
     Vue.use(VueImgix, {
       domain: 'test-domain.imgix.net',
       includeLibraryParam: false,
     });
 
-    const { buildUrlObject } = require('@/plugins/vue-imgix');
+    const { buildUrlObject } = require('@/plugins/imgix-vue');
     expect(buildUrlObject('/test-image.jpg')).toMatchObject({
       src: expect.not.stringMatching(/ixlib/),
       srcset: expect.not.stringMatching(/ixlib/),
@@ -35,7 +35,7 @@ describe('Initialization', () => {
 
   it(`includes imgixParams set during initialization in the generated srcs`, () => {
     const Vue = require('vue');
-    const VueImgix = require('@/plugins/vue-imgix');
+    const VueImgix = require('@/plugins/imgix-vue');
     Vue.use(VueImgix, {
       domain: 'test-domain.imgix.net',
       defaultIxParams: {
@@ -43,7 +43,7 @@ describe('Initialization', () => {
       },
     });
 
-    const { buildUrlObject } = require('@/plugins/vue-imgix');
+    const { buildUrlObject } = require('@/plugins/imgix-vue');
     expect(buildUrlObject('/test-image.jpg')).toMatchObject({
       src: expect.stringMatching(/auto=format/),
       srcset: expect.stringMatching(/auto=format/),
